Placed u.fl coax connector insight a hole of the battery pastic frame.
I replaced the 220pF block cap in front of the antenna with a cap jumper to switch between the antenna and the u.fl coax connector.
To create some space for the coax connector I modified the casio board footprint. At the footprint I moved and reshaped the spy-bi-wire pads.
Travis used clip connectors on one of the photos so I added some holes to to pads. If it is not possible to use clip connectors we could solder a 50 mil header or cable on to these pads (maybe even using a pogo pin adapter?).
The u.fl connector is placed at the approximate position of the hole in the pastic frame (this might not fit since I am still missing the accurate dimentions of the pastic frame).
I added some markings to the top silk screen to visualize the jumper settings and resized/repositioned some component markings.

The routing of VCC on the ~RST pullup (R1) is not that nice anymore.